+++ dlt/helpers/dbt/__init__.py
@@ -1,11 +1,9 @@
-from typing import Sequence
-
 from dlt.common.runners.venv import Venv
 from dlt.common.destination import DestinationClientDwhConfiguration
 from dlt.common.configuration.specs import CredentialsWithDefault
 from dlt.common.typing import TSecretValue
 
-from dlt.dbt.runner import create_runner, DBTPackageRunner
+from dlt.helpers.dbt.runner import create_runner, DBTPackageRunner
 
 
 def _default_profile_name(credentials: DestinationClientDwhConfiguration,) -> str:
