+++ tests/dbt_runner/local/test_dbt_utils.py
@@ -8,7 +8,7 @@ from dlt.common.runners.synth_pickle import decode_obj
 
 from dlt.common.storages import FileStorage
 from dlt.common.utils import uniq_id
-from dlt.dbt.dbt_utils import DBTProcessingError, initialize_dbt_logging, run_dbt_command, is_incremental_schema_out_of_sync_error
+from dlt.helpers.dbt.dbt_utils import DBTProcessingError, initialize_dbt_logging, run_dbt_command, is_incremental_schema_out_of_sync_error
 from tests.utils import test_storage, preserve_environ
 from tests.dbt_runner.utils import clone_jaffle_repo, load_test_case
